+++ packages/hoprd/src/index.ts
@@ -413,7 +413,7 @@ async function main() {
 
   try {
     logs.log(`This is HOPRd version ${version}`)
-    if (on_avado) logs.log('This node appears to be running on an AVADO')
+    if (on_avado) logs.log('This node appears to be running on an AVADO/Dappnode')
 
     // 1. Find or create an identity
     const peerId = await getIdentity({