+++ CHANGELOG.md
@@ -41,6 +41,7 @@
   - command parsing has been overhauled to support more complex commands
   - improved user experience with more consistent messages
 - Warn when manual path selection contains duplicate adjacent entries
+- Correctly recognize Avado/Dappnode private subnet
 
 # Breaking changes